requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
slowapi==0.1.4
openai==1.3.5
google-generativeai==0.8.5
python-dotenv==1.0.0
pydub # Requires FFmpeg or avconv installed on the system
numpy==1.24.3
torch==2.1.0
torchaudio==2.1.0
speechbrain==0.5.15
pyannote.audio==3.1.1
librosa==0.10.1
soundfile==0.12.1
whisper @ git+https://github.com/openai/whisper.git 